Refactor renderer.js into focused modules
Split the 1,678-line renderer.js into 5 focused modules:
- themes.js: ANSI codes, color math, theme definitions, thought bubbles
- animations.js: eye and mouth animation functions
- particles.js: ParticleSystem class (10 visual effect styles)
- face.js: ClaudeFace class (single face mode)
- grid.js: MiniFace + FaceGrid classes (grid mode)

renderer.js is now a slim entry point (~230 lines) with runtime loops,
PID guard, and re-exports for backward compatibility. Test imports
updated to reference new modules directly. All 223 tests pass.
